+++ lib/Fhaculty/Graph/Algorithm/TravelingSalesmanProblem/Result.php
@@ -6,65 +6,31 @@ use Fhaculty\Graph\Walk;
 use Fhaculty\Graph\Vertex;
 use Fhaculty\Graph\Edge\Base as Edge;
 use Fhaculty\Graph\Set\Edges;
-use Fhaculty\Graph\Algorithm\Base as AlgorithmBase;
+use Fhaculty\Graph\Set\EdgesAggregate;
 
-abstract class Result extends AlgorithmBase
+interface Result extends EdgesAggregate
 {
     /**
      * get resulting graph with the (first) best circle of edges connecting all vertices
      *
      * @throws Exception on error
      * @return Graph
-     * @uses Tsp::getGraph()
-     * @uses AlgorithmTsp::getEdges()
-     * @uses Graph::createGraphCloneEdges()
      */
-    public function createGraph()
-    {
-        return $this->getGraph()->createGraphCloneEdges($this->getEdges());
-    }
-
-    /**
-     * get graph this algorithm operates on
-     *
-     * @return Graph
-     */
-    abstract protected function getGraph();
-
-    /**
-     * get start vertex this algorithm starts on
-     *
-     * @return Vertex
-     */
-    abstract protected function getVertexStart();
+    public function createGraph();
 
     /**
      * get (first) best circle connecting all vertices
      *
      * @return Walk
-     * @uses AlgorithmTsp::getEdges()
-     * @uses AlgorithmTsp::getVertexStart()
-     * @uses Walk::factoryCycleFromEdges()
      */
-    public function getCycle()
-    {
-        return Walk::factoryCycleFromEdges($this->getEdges(), $this->getVertexStart());
-    }
+    public function getCycle();
 
-    public function getWeight()
-    {
-        $weight = 0;
-        foreach ($this->getEdges() as $edge) {
-            $weight += $edge->getWeight();
-        }
-
-        return $weight;
-    }
+    public function getWeight();
 
     /**
      * get array of edges connecting all vertices in a circle
      *
      * @return Edges
      */
-    abstract public function getEdges();
+    // public function getEdges();
 }
